[ADD] long->int type변환

+++ src/main/java/com/ssafy/sgdc/category/CategoryRepo.java
@@ -3,8 +3,8 @@ package com.ssafy.sgdc.category;
 import com.ssafy.sgdc.domain.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepo extends JpaRepository<Category, Long> {
+public interface CategoryRepo extends JpaRepository<Category, Integer> {
 
-    Category findByCategoryId(Long categoryId);
+    Category findByCategoryId(int categoryId);
 
 }
